pyproject.toml
[project]
name = "arboles-madrid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "folium>=0.20.0",
    "httpx>=0.28.1",
    "networkx>=3.6.1",
    "numpy>=2.4.2",
    "openpyxl>=3.1.5",
    "osmnx>=2.1.0",
    "pandas>=3.0.1",
    "pyproj>=3.7.2",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "shapely>=2.1.2",
    "tqdm>=4.67.3",
    "uvicorn[standard]>=0.41.0",
]
